fix: four defects the date_trunc work uncovered
None of these belong to the date_trunc change; they are pre-existing bugs it
ran into, each reachable from plain SQL, and each fixed here so the branch is
green. All four are worth carrying upstream.

`StrfTimeFormat::WritePadded` writes the trailing three digits of an odd-width
field, subtracts three from the padding, and returns a pointer three bytes
short of what it wrote. `%n` is the only odd-width caller, so a specifier after
`%n` overwrote the last three nanosecond digits and left uninitialised bytes
behind, which fails on insert as invalid unicode.

Comparison simplification reads the time of day off the constant to decide
whether it can push it through a cast, and Timestamp::GetTime throws for an
infinite timestamp, so `last_day(ts)::TIMESTAMP <= TIMESTAMP 'infinity'` failed
where the same query with the expression rewriter disabled answered fine. An
infinite timestamp has no time of day, so the simplification declines for it.

json_serialize_plan resolved the column bindings before the types. An
operator's cached types can be stale after optimization, which is why the
physical planner resolves the types first; doing it the other way round meant
the resolver met an operator claiming more columns than it had bindings for.
Deduplicating the sum and count that `avg` expands into was enough to trigger
it.

The ICU timestamp difference reads the end value's microseconds as an unsigned
remainder, which is negative before 1970 and wraps; the unsigned sum cancels
the wrap again, but the result went through a checked cast, so subtracting two
1937 timestamps with sub-millisecond parts answered correctly in a release
build and threw in an assert one. Flooring the remainder would change the
answer, since the borrow below it and the calendar field differences are
written for the truncated form, so the cast says what the arithmetic means.

(cherry picked from commit cd7da67d5ae76a334e235c0f823cdb0496496323)

+++ extension/icu/icu-dateadd.cpp
@@ -171,7 +171,8 @@ interval_t ICUCalendarSub::Operation(timestamp_tz_t end_date, timestamp_tz_t sta
 	auto min_diff = SubtractField(calendar, UCAL_MINUTE, end_date);
 	auto sec_diff = SubtractField(calendar, UCAL_SECOND, end_date);
 	auto ms_diff = SubtractField(calendar, UCAL_MILLISECOND, end_date);
-	auto micros_diff = UnsafeNumericCast<int32_t>(ms_diff * Interval::MICROS_PER_MSEC + (end_micros - start_micros));
+	auto micros_diff = UnsafeNumericCast<int32_t>(ms_diff * Interval::MICROS_PER_MSEC +
+	                                             (static_cast<int64_t>(end_micros) - static_cast<int64_t>(start_micros)));
 	result.micros = Time::FromTime(hour_diff, min_diff, sec_diff, micros_diff).micros;
 
 	return result;
@@ -210,7 +211,8 @@ interval_t ICUCalendarAge::Operation(timestamp_tz_t end_date, timestamp_tz_t sta
 	auto min_diff = SubtractField(calendar, UCAL_MINUTE, end_date);
 	auto sec_diff = SubtractField(calendar, UCAL_SECOND, end_date);
 	auto ms_diff = SubtractField(calendar, UCAL_MILLISECOND, end_date);
-	auto micros_diff = UnsafeNumericCast<int32_t>(ms_diff * Interval::MICROS_PER_MSEC + (end_micros - start_micros));
+	auto micros_diff = UnsafeNumericCast<int32_t>(ms_diff * Interval::MICROS_PER_MSEC +
+	                                             (static_cast<int64_t>(end_micros) - static_cast<int64_t>(start_micros)));
 	result.micros = Time::FromTime(hour_diff, min_diff, sec_diff, micros_diff).micros;
 
 	return result;
